refactor: jdbc 관련 파일 제거
