+++ src/main/java/gift/domain/product/Product.java
@@ -6,7 +6,6 @@ import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import org.hibernate.validator.constraints.UniqueElements;
 
 @Entity
 public class Product {
