--- src/main/java/gift/domain/product/repository/ProductRepository.java
@@ -1,6 +1,0 @@
-package gift.domain.product.repository;
-
-
-public interface ProductRepository {
-
-}
