--- src/main/java/gift/domain/user/repository/UserRepository.java
@@ -1,5 +1,0 @@
-package gift.domain.user.repository;
-
-public interface UserRepository {
-
-}
