+++ src/main/java/gift/global/controller/HomeController.java
@@ -10,7 +10,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
-@RequestMapping( "/api")
+@RequestMapping( "/")
 public class HomeController {
 
     private final ProductService productService;
