+++ src/test/java/gift/JpaDataTest/CartItemRepositoryTest.java
@@ -5,9 +5,9 @@ import static org.assertj.core.api.Assertions.*;
 import gift.domain.cart.CartItem;
 import gift.domain.cart.JpaCartItemRepository;
 import gift.domain.product.Product;
-import gift.domain.product.repository.JpaProductRepository;
+import gift.domain.product.JpaProductRepository;
 import gift.domain.user.User;
-import gift.domain.user.repository.JpaUserRepository;
+import gift.domain.user.JpaUserRepository;
 import java.util.List;
 import jdk.jfr.Description;
 import org.junit.jupiter.api.Test;
